Source reorganization - database, encryption, networking
