+++ src/database/databasemanager.h
@@ -17,34 +17,22 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-#ifndef FS_STATUS_H_8B28B354D65B4C0483E37AD1CA316EB4
-#define FS_STATUS_H_8B28B354D65B4C0483E37AD1CA316EB4
+#ifndef FS_DATABASEMANAGER_H_2B75821C555E4D1D83E32B20D683217C
+#define FS_DATABASEMANAGER_H_2B75821C555E4D1D83E32B20D683217C
+#include "database/database.h"
 
-#include "networkmessage.h"
-#include "protocol.h"
-
-class ProtocolStatus final : public Protocol
+class DatabaseManager
 {
 	public:
-		// static protocol information
-		enum {server_sends_first = false};
-		enum {protocol_identifier = 0xFF};
-		enum {use_checksum = false};
-		static const char* protocol_name() {
-			return "status protocol";
-		}
+		static bool tableExists(const std::string& table);
 
-		explicit ProtocolStatus(Connection_ptr connection) : Protocol(connection) {}
+		static int32_t getDatabaseVersion();
+		static bool isDatabaseSetup();
 
-		void onRecvFirstMessage(NetworkMessage& msg) final;
+		static bool optimizeTables();
+		static void updateDatabase();
 
-		void sendStatusString();
-		void sendInfo(uint16_t requestedInfo, const std::string& characterName);
-
-		static const uint64_t start;
-
-	protected:
-		static std::map<uint32_t, int64_t> ipConnectMap;
+		static bool getDatabaseConfig(const std::string& config, int32_t& value);
+		static void registerDatabaseConfig(const std::string& config, int32_t value);
 };
-
 #endif
