+++ src/database/databasetasks.h
@@ -22,7 +22,7 @@
 
 #include <condition_variable>
 #include "thread_holder_base.h"
-#include "database.h"
+#include "database/database.h"
 #include "enums.h"
 
 struct DatabaseTask {
