+++ src/game.cpp
@@ -26,7 +26,7 @@
 #include "configmanager.h"
 #include "creature.h"
 #include "creatureevent.h"
-#include "databasetasks.h"
+#include "database/databasetasks.h"
 #include "events.h"
 #include "game.h"
 #include "globalevent.h"
@@ -36,7 +36,7 @@
 #include "monster.h"
 #include "movement.h"
 #include "scheduler.h"
-#include "server.h"
+#include "networking/server.h"
 #include "spells.h"
 #include "talkaction.h"
 #include "weapons.h"
