+++ src/iomarket.h
@@ -21,7 +21,7 @@
 #define FS_IOMARKET_H_B981E52C218C42D3B9EF726EBF0E92C9
 
 #include "enums.h"
-#include "database.h"
+#include "database/database.h"
 
 class IOMarket
 {
