+++ src/luascript.h
@@ -32,7 +32,7 @@
 #endif
 #endif
 
-#include "database.h"
+#include "database/database.h"
 #include "enums.h"
 #include "position.h"
