+++ src/networking/networkmessage.h
@@ -23,10 +23,7 @@
 #include "const.h"
 
 class Item;
-class Creature;
-class Player;
 struct Position;
-class RSA;
 
 class NetworkMessage
 {
