+++ src/networking/protocols/protocol.h
@@ -20,7 +20,7 @@
 #ifndef FS_PROTOCOL_H_D71405071ACF4137A4B1203899DE80E1
 #define FS_PROTOCOL_H_D71405071ACF4137A4B1203899DE80E1
 
-#include "connection.h"
+#include "networking/connection.h"
 
 class Protocol : public std::enable_shared_from_this<Protocol>
 {
