+++ src/networking/server.h
@@ -20,7 +20,7 @@
 #ifndef FS_SERVER_H_984DA68ABF744127850F90CC710F281B
 #define FS_SERVER_H_984DA68ABF744127850F90CC710F281B
 
-#include "connection.h"
+#include "networking/connection.h"
 #include "signals.h"
 #include <memory>
