+++ src/quests.h
@@ -21,7 +21,7 @@
 #define FS_QUESTS_H_16E44051F23547BE8097F8EA9FCAACA0
 
 #include "player.h"
-#include "networkmessage.h"
+#include "networking/networkmessage.h"
 
 class Mission;
 class Quest;
